REmoved the POM processing stuff from the Antikythera class
