+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/SpringEvaluator.java
@@ -340,6 +340,7 @@ public class SpringEvaluator extends ControlFlowEvaluator {
                 }
             }
         } catch (Exception e) {
+            e.printStackTrace();
             handleApplicationException(e);
         }
     }
