+++ src/main/java/sa/com/cloudsolutions/antikythera/generator/Antikythera.java
@@ -1,9 +1,5 @@
 package sa.com.cloudsolutions.antikythera.generator;
 
-import org.apache.maven.model.Dependency;
-import org.apache.maven.model.Model;
-import org.apache.maven.model.io.xpp3.MavenXpp3Reader;
-import org.apache.maven.model.io.xpp3.MavenXpp3Writer;
 import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,38 +7,30 @@ import sa.com.cloudsolutions.antikythera.configuration.Settings;
 import sa.com.cloudsolutions.antikythera.exception.AntikytheraException;
 import sa.com.cloudsolutions.antikythera.exception.EvaluatorException;
 import sa.com.cloudsolutions.antikythera.parser.AbstractCompiler;
+import sa.com.cloudsolutions.antikythera.parser.MavenHelper;
 import sa.com.cloudsolutions.antikythera.parser.RestControllerParser;
 import sa.com.cloudsolutions.antikythera.parser.ServicesParser;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.channels.Channels;
-import java.nio.channels.FileChannel;
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.util.ArrayList;
+
 import java.util.Collection;
-import java.util.List;
-import java.util.Properties;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class Antikythera {
-    public static final String POM_XML = "pom.xml";
+
     public static final String SRC = "src";
     private static final Logger logger = LoggerFactory.getLogger(Antikythera.class);
     private static final String PACKAGE_PATH = "src/main/java/sa/com/cloudsolutions/antikythera";
     private static Antikythera instance;
     private final Collection<String> controllers;
     private final Collection<String> services;
-    private Model pomModel;
+    private static MavenHelper mavenHelper;
+
 
     private Antikythera() {
         controllers = Settings.getPropertyList(Settings.CONTROLLERS, String.class);
@@ -54,7 +42,9 @@ public class Antikythera {
             try {
                 Settings.loadConfigMap();
                 instance = new Antikythera();
-                instance.readPomFile();
+                mavenHelper = new MavenHelper();
+                mavenHelper.readPomFile();
+                mavenHelper.buildJarPaths();
             } catch (IOException e) {
                 throw new AntikytheraException("Failed to initialize Antikythera", e);
             } catch (XmlPullParserException xe) {
@@ -81,7 +71,6 @@ public class Antikythera {
                 Files.copy(sourcePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
             }
         }
-
     }
 
     public static void main(String[] args) throws IOException, XmlPullParserException, EvaluatorException {
@@ -97,111 +86,11 @@ public class Antikythera {
         antk.generateUnitTests();
     }
 
-    /**
-     * Copy a template file to the specified path
-     *
-     * @param filename the name of the template file to copy
-     * @param subPath  the path components
-     * @throws IOException thrown if the copy operation failed
-     */
-    private void copyTemplate(String filename, String... subPath) throws IOException {
-        Path destinationPath = Path.of(Settings.getOutputPath(), subPath);     // Path where template file should be copied into
-        Files.createDirectories(destinationPath);
-        try (InputStream sourceStream = getClass().getClassLoader().getResourceAsStream("templates/" + filename);
-             FileOutputStream destStream = new FileOutputStream(destinationPath + File.separator + filename);
-             FileChannel destChannel = destStream.getChannel()) {
-            if (sourceStream == null) {
-                throw new IOException("Template file not found");
-            }
-            destChannel.transferFrom(Channels.newChannel(sourceStream), 0, Long.MAX_VALUE);
-        }
-    }
-
-    /**
-     * Copy the pom.xml file to the specified path, injecting the dependencies as needed.
-     * <p>
-     * The configuration file needs to have a dependencies section that provides the list of
-     * `artifactids` that need to be included in the output pom file. If these dependencies
-     * require any variables, those are copied as well.
-     * <p>
-     * The primary source file is the file from the template folder. The dependencies are
-     * supposed to be listed in the pom file of the application under test.
-     *
-     * @throws IOException            if the POM file cannot be copied
-     * @throws XmlPullParserException if the POM file cannot be converted to an XML Tree
-     */
-    public void copyPom() throws IOException, XmlPullParserException {
-        String[] dependencies = Settings.getArtifacts();
-        if (dependencies.length == 0) {
-            copyTemplate(POM_XML);
-        } else {
-            Path destinationPath = Path.of(Settings.getOutputPath(), POM_XML);
-
-            MavenXpp3Reader reader = new MavenXpp3Reader();
-            Model templateModel = reader.read(getClass().getClassLoader().getResourceAsStream("templates/pom.xml"));
-
-            List<Dependency> srcDependencies = pomModel.getDependencies();
-            for (String dep : dependencies) {
-                for (Dependency dependency : srcDependencies) {
-                    if (dependency.getArtifactId().equals(dep)) {
-                        templateModel.addDependency(dependency);
-                        copyDependencyProperties(pomModel, templateModel, dependency);
-                    }
-                }
-            }
-
-            MavenXpp3Writer writer = new MavenXpp3Writer();
-            writer.write(new FileWriter(destinationPath.toFile()), templateModel);
-        }
-    }
-
-    private void readPomFile() throws IOException, XmlPullParserException {
-        MavenXpp3Reader reader = new MavenXpp3Reader();
-        String basePath = Settings.getBasePath();
-        Path p = null;
-        if (basePath.contains("src/main/java")) {
-            p = Paths.get(basePath.replace("/src/main/java", ""), POM_XML);
-        } else if (basePath.contains("src/test/java")) {
-            p = Paths.get(basePath.replace("/src/test/java", ""), POM_XML);
-        } else {
-            p = Paths.get(basePath, POM_XML);
-        }
-
-        pomModel = reader.read(new FileReader(p.toFile()));
-    }
-
-    /**
-     * Copy the properties of the dependency from the source model to the template model
-     *
-     * @param srcModel      the pom file model from the application under test
-     * @param templateModel from the template folder
-     * @param dependency    the dependency that may or may not have a property
-     */
-    private void copyDependencyProperties(Model srcModel, Model templateModel, Dependency dependency) {
-        Properties srcProperties = srcModel.getProperties();
-        Properties templateProperties = templateModel.getProperties();
-        Pattern pattern = Pattern.compile("\\$\\{(.+?)\\}");
-
-        // Check all fields of the dependency for property references
-        String[] fields = {dependency.getGroupId(), dependency.getArtifactId(), dependency.getVersion(), dependency.getScope()};
-        for (String field : fields) {
-            if (field != null) {
-                Matcher matcher = pattern.matcher(field);
-                while (matcher.find()) {
-                    String propertyName = matcher.group(1);
-                    if (srcProperties.containsKey(propertyName) && !templateProperties.containsKey(propertyName)) {
-                        templateProperties.setProperty(propertyName, srcProperties.getProperty(propertyName));
-                    }
-                }
-            }
-        }
-    }
-
     private void copyBaseFiles(String outputPath) throws IOException, XmlPullParserException {
         String testPath = PACKAGE_PATH.replace("main", "test");
-        copyPom();
-        copyTemplate("TestHelper.java", testPath, "base");
-        copyTemplate("Configurations.java", testPath, "configurations");
+        mavenHelper.copyPom();
+        mavenHelper.copyTemplate("TestHelper.java", testPath, "base");
+        mavenHelper.copyTemplate("Configurations.java", testPath, "configurations");
 
         Path pathToCopy = Paths.get(outputPath, SRC, "test", "resources");
         Files.createDirectories(pathToCopy);
@@ -249,88 +138,6 @@ public class Antikythera {
         }
     }
 
-    public String[] getJarPaths() {
-        List<String> jarPaths = new ArrayList<>();
-
-        if (pomModel != null) {
-            List<Dependency> dependencies = pomModel.getDependencies();
-            Properties properties = pomModel.getProperties();
-            Settings.getProperty("variables.m2_folder", String.class)
-                    .ifPresent(m2 -> {
-                        for (Dependency dependency : dependencies) {
-                            addJarPath(dependency, properties, m2, jarPaths);
-                        }
-                    });
-        }
-        return jarPaths.toArray(new String[0]);
-    }
-
-    private void addJarPath(Dependency dependency, Properties properties, String m2, List<String> jarPaths) {
-        String groupIdPath = dependency.getGroupId().replace('.', '/');
-        String artifactId = dependency.getArtifactId();
-        String version = dependency.getVersion();
-
-        // Handle property variables in version
-        if (version != null && version.startsWith("${") && version.endsWith("}")) {
-            String propertyName = version.substring(2, version.length() - 1);
-            version = properties.getProperty(propertyName);
-        }
-
-        if (version == null || version.isEmpty()) {
-            version = findLatestVersion(groupIdPath, artifactId, m2);
-        }
-
-        if (version != null) {
-            Path p = Paths.get(m2, groupIdPath, artifactId, version,
-                    artifactId + "-" + version + ".jar");
-            if (Files.exists(p)) {
-                jarPaths.add(p.toString());
-            } else {
-                logger.warn("Jar not found: {}", p);
-            }
-        }
-    }
-
-    private String findLatestVersion(String groupIdPath, String artifactId, String m2) {
-        Path artifactPath = Paths.get(m2, groupIdPath, artifactId);
-
-        if (!Files.exists(artifactPath) || !Files.isDirectory(artifactPath)) {
-            return null;
-        }
-
-        try (DirectoryStream<Path> stream = Files.newDirectoryStream(artifactPath)) {
-            List<String> versions = new ArrayList<>();
-            for (Path path : stream) {
-                if (Files.isDirectory(path)) {
-                    String version = path.getFileName().toString();
-                    Path p = Paths.get(m2, groupIdPath, artifactId, version, artifactId + "-" + version + ".jar");
-                    if (! (version.startsWith("${") || version.equals("unknown")) && Files.exists(p)) {
-                        versions.add(version);
-                    }
-                }
-            }
-
-            return versions.stream().max(this::compareVersions).orElse(null);
-        } catch (IOException e) {
-            return null;
-        }
-    }
-
-
-    private int compareVersions(String v1, String v2) {
-        String[] parts1 = v1.split("\\.");
-        String[] parts2 = v2.split("\\.");
-
-        int length = Math.max(parts1.length, parts2.length);
-        for (int i = 0; i < length; i++) {
-            int part1 = i < parts1.length ? Integer.parseInt(parts1[i]) : 0;
-            int part2 = i < parts2.length ? Integer.parseInt(parts2[i]) : 0;
-            if (part1 != part2) {
-                return Integer.compare(part1, part2);
-            }
-        }
-        return 0;
-    }
     public void preProcess() throws IOException, XmlPullParserException {
         CopyUtils.createMavenProjectStructure(Settings.getBasePackage(), Settings.getOutputPath());
         copyBaseFiles(Settings.getOutputPath());
